requirements.txt
pandas
openpyxl
matplotlib
google-generativeai
Pillow
tk
customtkinter
python-dotenv